% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchSigsAndRelabel.R
\name{MatchSigsAndRelabel}
\alias{MatchSigsAndRelabel}
\title{Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.}
\usage{
MatchSigsAndRelabel(ex.sigs, gt.sigs, exposure = NULL)
}
\arguments{
\item{ex.sigs}{Newly extracted signatures to be compared to gt.sigs}

\item{gt.sigs}{"Ground truth" signatures.}

\item{exposure}{If \code{NULL}, then match
\code{ex.sigs} against all signatures in \code{gt.sigs}.
Otherwise this should be ground-truth exposures used generate the
synthetic spectra from which \code{ex.sigs} were extracted.
In this case we do not
match to ground-truth signatures to that were not in the ground
truth exposure.}
}
\value{
A list with the elements \code{averCosSim}, \code{match1},
\code{match2} as for \code{SigSetSimilarity}, with \code{match1}
 being matches for the the extracted signatures (\code{ex.sigs})
 and \code{match2} being the
 matches for the ground truth signatures (\code{gt.sigs}). The return list
 also echos the input arguments \code{ex.sigs} and \code{gt.sigs}.
}
\description{
Run \code{MatchSigs2Directions}, then
plot its results and write them as .csv files.
}
\seealso{
Other signature matching functions: 
\code{\link{Match1Sig}()},
\code{\link{MatchSigs1Direction}()},
\code{\link{MatchSigs2Directions}()}
}
\concept{signature matching functions}
