% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewMatchSigsAndRelabel.R
\name{NewMatchSigsAndRelabel}
\alias{NewMatchSigsAndRelabel}
\title{A wrapper function around \code{\link[ICAMSxtra]{match_two_sig_sets()}}
and substitutes deprecate function \code{\link[ICAMSxtra]{MatchSigsAndRelabel()}},}
\usage{
NewMatchSigsAndRelabel(
  ex.sigs,
  gt.sigs,
  exposure = NULL,
  similarity.cutoff = 0.9
)
}
\arguments{
\item{ex.sigs}{Newly extracted signatures to be compared to \code{gt.sigs}.}

\item{gt.sigs}{"Ground truth" signatures.}

\item{exposure}{If \code{NULL}, then match
\code{ex.sigs} against all signatures in \code{gt.sigs}.
Otherwise this should be ground-truth exposures used generate the
synthetic spectra from which \code{ex.sigs} were extracted.
In this case we do not
match to ground-truth signatures that were not in the ground
truth exposure.}

\item{similarity.cutoff}{A ground-truth signature must have been
the best match of an extracted signature with a cosine
similarity \eqn{$ge$} this value to be considered a true positive.
Otherwise we consider the ground-truth signature to be a false
negative.}
}
\value{
A list with the elements

* \code{averCosSim} The average of the cosine similarities
   between each signature in
   \code{ex.sigs} and its closest match in \code{gt.sigs}
   and the closest match
   between each signature in \code{gt.sigs}
   and its closest match in \code{ex.sigs}.
   This may not be what you want. Often one wants
   the average of the cosine similarities of the true
   positives to their matching ground-truth signatures.

* \code{match1} The match from extracted signatures to ground truth
        signatures. Associated with each extracted signature is
        a ground truth signature with best cosine similarity.
        TODO: Ties?

* \code{match2} The match from ground truth signatures to extracted
        signatures. Associated with each extracted signature is
        a ground truth signature with best cosine similarity.
        TODO: Ties?

* \code{extracted.with.no.best.match}
* \code{ground.truth.with.no.best.match}
* \code{ex.sigs} Echo input argument
* \code{gt.sigs} Echo input argument
* \code{gt.mean.cos.sim}
}
\description{
A wrapper function around \code{\link[ICAMSxtra]{match_two_sig_sets()}}
and substitutes deprecate function \code{\link[ICAMSxtra]{MatchSigsAndRelabel()}},
}
\examples{
gt.sigs <- matrix(c(1, 3,   4, 1, 2, 4), ncol = 2)
ex.sigs <- matrix(c(1, 3.1, 4, 5, 1, 1, 1, 2.8, 4), ncol = 3)
colnames(gt.sigs) <- c("gt.1", "gt.2")
colnames(ex.sigs) <- c("ex.1", "ex.2", "ex.3")
tout <- MatchSigsAndRelabel(gt.sigs = gt.sigs, ex.sigs = ex.sigs)
tout

}
