% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAndAnalyzeExposures.R
\name{ReadAndAnalyzeExposures}
\alias{ReadAndAnalyzeExposures}
\title{Assess how well inferred exposures match input exposures}
\usage{
ReadAndAnalyzeExposures(
  extracted.sigs,
  reference.sigs,
  inferred.exp.path,
  ground.truth.exposures
)
}
\arguments{
\item{extracted.sigs}{Path to file containing the extracted signature profiles.}

\item{reference.sigs}{File containing signature profiles from which the
synthetic data were generated.}

\item{inferred.exp.path}{File containing mutation counts (exposures)
of synthetic tumors which are inferred to extracted or input signatures.}

\item{ground.truth.exposures}{File containing the exposures from which
the synthetic catalogs were generated.  This file is used to restrict
assessment of signature exposures to only those signatures in
\code{reference.sigs} that were actually represented in the exposures.}
}
\value{
A \code{\link{data.frame}} recording:

\code{Ground.truth.exposure}: sum of ground truth exposures of
all tumors to all reference signatures.

\code{Inferred.exposure}: sum of inferred exposures of
all tumors to all reference signatures.
Here, inferred exposure of a tumor to a ground-truth
signature equals to the sum of the exposures of this tumor
to all extracted signatures which are most similar to
a reference signature.
If there is no extracted signature resembling an ground-truth
signature, the inferred exposure of this ground-truth
signature will be \code{0}.

\code{Absolute.difference}: sum of absolute difference between
ground-truth exposure and inferred exposure of all tumors
to all reference signatures.
}
\description{
We assume that in many cases attribution programs will be run
outside of R on file inputs and will generate fill outputs.
}
\details{
Generates output files by calling
\code{\link[mSigTools]{TP_FP_FN_avg_sim}}
}
