% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeMultiToolsMultiDatasets.R
\name{SummarizeMultiToolsMultiDatasets}
\alias{SummarizeMultiToolsMultiDatasets}
\title{Summarize results for multiple datasets, by different computational approaches.}
\usage{
SummarizeMultiToolsMultiDatasets(
  toolSummaryPaths,
  out.dir,
  display.datasetName = FALSE,
  sort.by.composite.extraction.measure = "descending",
  overwrite = FALSE
)
}
\arguments{
\item{toolSummaryPaths}{Paths of top-level dataset directories trees you want
to investigate.
E.g. "./S.0.1.Rsq.0.1"

Note: \code{OneToolSummary.RDa} are expected to be exist under
\code{toolSummaryPaths}.}

\item{out.dir}{Path of the output directory.}

\item{display.datasetName}{Whether to put the name of spectra datasets inside of
the csv outputs of summary tables.}

\item{sort.by.composite.extraction.measure}{Whether to re-order the computational
approaches on violin plots, based on the mean of composite measure.

\code{"descending"}: Put the computational approach with the highest mean composite
measure to the left, and arrange approaches in descending order.
\code{"ascending"}:  Put the computational approach with the lowest mean composite
measure to the left, and arrange approaches in ascending order.
Anything else: Keep the computational approaches in a smart alphabetical order embedded with numbers,
defined by \code{\link[gtools]{mixedsort}}.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results of mutational signature extraction and exposure inference
by multiple computational approaches on multiple datasets. Before running this
function, make sure the summary file for each single data set
 \code{toolSummaryPaths}/\code{OneToolSummary.Rda} exists.
}
\details{
\code{OneToolSummary.Rda} is generated by \code{\link{SummarizeOneToolMultiDatasets}}).
}
