% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeOneToolMultiDatasets.R
\name{SummarizeOneToolMultiDatasets}
\alias{SummarizeOneToolMultiDatasets}
\title{Combine results for multiple datasets, from one computational approaches.}
\usage{
SummarizeOneToolMultiDatasets(
  dataset.dirs,
  datasetGroup,
  datasetGroupName,
  datasetSubGroup = NULL,
  datasetSubGroupName = NULL,
  toolName,
  tool.dirname,
  out.dir,
  overwrite = FALSE
)
}
\arguments{
\item{dataset.dirs}{Paths of top-level dataset directories trees you want
to investigate. Should be folders under \code{tool.dirname}.
E.g. S.0.1.Rsq.0.1}

\item{datasetGroup}{Numeric or character vector specifying the group
each dataset belong to.
E.g. For SBS1-SBS5 correlated datasets, we can consider slope
(SBS1:SBS5 count ratio) as the group:
\code{rep(c(0.1,0.5,1,2,5,10),each = 4)}

The value is set to \code{Default} if unspecified.}

\item{datasetGroupName}{Meaning of all datasetGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider
\code{"SBS1:SBS5 mutation count ratio"}
as what \code{datasetGroup} is referring to.}

\item{datasetSubGroup}{Numeric or character vector differentiating
datasets within each group.
E.g. For SBS1-SBS5 correlated datasets, we can consider Pearson's R^2
as the subgroup:
\code{rep(c(0.1,0.2,0.3,0.6),times = 5)}}

\item{datasetSubGroupName}{Meaning of all \code{datasetSubGroup}.
E.g. For SBS1-SBS5 correlated datasets, we can consider \code{"Pearson's R squared"}
as what \code{datasetSubGroup} is referring to.}

\item{toolName}{Name of computational approach to be investigated
(e.g. "SigProExtractor")}

\item{tool.dirname}{Combining top-level directory
(e.g. "3a.Original_output_K_unspecified") and second-level
directory containing outputs and summaries of one computational approach
to be investigated (e.g. "SigProExtractor" or "SigProExtractor.results").

One example: "3a.Original_output_K_unspecified/SigProExtractor"

Note: this function expects file \code{multiRun.RDa} generated by
\code{SummarizeMultiRuns} under \code{tool.dirname}/\code{dataset.dirs}}

\item{out.dir}{Path of the output directory.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results from each computational approach in \code{third.level.dir}/\code{tool.dirnames}
(generated by \code{\link{SummarizeMultiRuns}}),
combine them into \code{third.level.dir}.
}
