% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSP.R
\name{SummarizeSigOneSigProExtractorSubdir}
\alias{SummarizeSigOneSigProExtractorSubdir}
\title{Assess/evaluate results from SigProExtractor
SigProfiler-python de novo extraction and attribution package.
Assessment is restricted to v0.0.5.43+,
because different version has different folder structure.}
\usage{
SummarizeSigOneSigProExtractorSubdir(
  run.dir,
  ground.truth.exposure.dir = paste0(run.dir, "/../../../"),
  summarize.exp = TRUE,
  overwrite = FALSE,
  hierarchy = FALSE
)
}
\arguments{
\item{run.dir}{A directory which contains output of SigProExtractor in one
run on a specific dataset, possibly with a specified seed. E.g.
\code{2b.Full_output_K_as_2/SigProExtractor.results/S.0.1.Rsq.0.1/seed.1/}.

This code depends on a conventional directory structure documented
in \code{NEWS.md}.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures. Should
contain a file named \code{ground.truth.syn.exposures.csv}.
In PCAWG paper: \code{run.dir/../../}
In SBS1-SBS5 paper: \code{0.Input_datasets/S.0.1.Rsq.0.1/}}

\item{summarize.exp}{Whether to summarize exposures when the file specified
by \code{inferred.exp.path} exists.}

\item{overwrite}{If TRUE overwrite and files in existing \code{run.dir/summary} folder.}

\item{hierarchy}{Whether the user have enabled hierarchy = True when running SigProExtractor.
specifying True or False into SigProExtractor will cause the program
to generate different folder structure.}
}
\description{
Assess/evaluate results from SigProExtractor
SigProfiler-python de novo extraction and attribution package.
Assessment is restricted to v0.0.5.43+,
because different version has different folder structure.
}
