% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeSigOneSubdir.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate one result sub-folder from a computational approach}
\usage{
SummarizeSigOneSubdir(
  run.dir,
  ground.truth.exposure.dir,
  extracted.sigs.path,
  inferred.exp.path = NULL,
  summarize.exp = TRUE,
  overwrite = FALSE,
  summary.folder.name = "summary",
  export.Manhattan.each.spectrum = FALSE
)
}
\arguments{
\item{run.dir}{A directory which contains output of computational approach
in one run on a specific dataset, possibly with a specified seed. E.g.
\code{2b.Full_output_K_as_2/hdp.results/S.0.1.Rsq.0.1/seed.1/}.

This code depends on a conventional directory structure documented
in \code{NEWS.md}.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
Should contain a file named \code{ground.truth.syn.exposures.csv}.
In PCAWG paper: \code{run.dir/../../}
In SBS1-SBS5 paper: \code{0.Input_datasets/S.0.1.Rsq.0.1/}}

\item{extracted.sigs.path}{Path to extracted sigs file.
For most computational approaches, it's in:
\code{<run.dir>/extracted.signatures.csv}.

For SigProExtractor, it's in:
\code{<run.dir>/<catalog.type>/Selected_Solution/De_Novo_Solution/De_Novo_Solution_Signatures_<catalog.type>.txt}
, and is converted to
\code{<run.dir>/extracted.signatures.csv} by
wrapper function \code{\link{SummarizeSigProExtractor}}.

For SignatureAnalyzer, it's in:
 \code{<run.dir>/sa.output.sigs.csv}

For EMu, helmsman.NMF, MultiModalMuSig.LDA and MultiModalMuSig.CTM,
their extracted signature files are not standard \code{ICAMS} catalog format.
These non-standard extracted signatures are converted to
\code{extracted.signatures.csv}, a standard \code{ICAMS} catalog
by their respective wrapper functions: \itemize{
\item \code{\link{ReadEMuCatalog}},
\item \code{\link{helmsmanCatalog2ICAMS}}, and
\item \code{\link{MMCatalog2ICAMS}}
}
They are dumped into \code{extracted.sigs.csv} same as most approaches.}

\item{inferred.exp.path}{Path to inferred exposures file.
For most computational approaches, it's in:
\code{<run.dir>/inferred.exposures.csv}.

For SigProExtractor, it's in
\code{<run.dir>/<catalog.type>/Selected_Solution/De_Novo_Solution/De_Novo_Solution_Activities_<catalog.type>.txt}
, and is converted to
\code{<run.dir>/inferred.exposures.csv} by
wrapper function \code{\link{SummarizeSigProExtractor}}.

For SignatureAnalyzer, it's in:
 \code{<run.dir>/sa.output.exp.csv}

For EMu, helmsman.NMF, MultiModalMuSig.LDA and MultiModalMuSig.CTM,
their extracted signature files are not standard \code{ICAMS} catalog format.
These non-standard extracted signatures are converted to
\code{extracted.signatures.csv}, a standard \code{mSigTools} exposures
by their respective wrapper functions: \itemize{
\item \code{\link{ReadEMuExposureFile}}
\item \code{\link{ReadhelmsmanExposure}}, and
\item \code{\link{ReadExposureMM}}
}
They are dumped into \code{inferred.exposures.csv} same as most approaches.}

\item{summarize.exp}{Whether to summarize exposures when the file specified
by \code{inferred.exp.path} exists.}

\item{overwrite}{If TRUE overwrite and files in existing \code{run.dir/summary} folder.}

\item{summary.folder.name}{The name of the folder containing summary results.
Usually, it equals to "summary".}

\item{export.Manhattan.each.spectrum}{Whether to export csv files for Manhattan
distance of each mutational spectrum.}
}
\description{
Note: For summarizing SigProExtractor or SignatureAnalyzer,
users should use SigProExtractor(SigProfiler-Python) v0.0.5.43+
and SignatureAnalyzer 2018-Apr-18.
}
\keyword{internal}
