% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelabelExSigs.R
\name{RelabelExSigs}
\alias{RelabelExSigs}
\title{Append most similar ground-truth signature and pairwise cosine similarity
to the name of each extracted signature in matrix of extracted signatures.}
\usage{
RelabelExSigs(sigAnalysis)
}
\arguments{
\item{sigAnalysis}{A list returned by function \code{\link{ReadAndAnalyzeSigs}},
at least including: \enumerate{
  \item \code{ex.sigs}: matrix of extracted signatures
  \item \code{sim.matrix}: full matrix between extracted signatures and
    ground-truth signatures
  \item \code{table}: matrix include pairs of true positive ground-truth
  signatures and true positive extracted signatures.
 }}
}
\value{
Matrix of extracted sigs, yet the names of signatures changed.
  New name: <old_name> (<name_of_most_similar_ground_truth_signature>
  cosine_similarity)
  e.g., Sig.A --> Sig.A (SBS1 0.998)
}
\description{
Append most similar ground-truth signature and pairwise cosine similarity
to the name of each extracted signature in matrix of extracted signatures.
}
