% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeMultiRuns}
\alias{SummarizeMultiRuns}
\title{Assess/evaluate multiple summarized runs for one dataset from one computational approach.}
\usage{
SummarizeMultiRuns(datasetName, toolName, tool.dir, run.names)
}
\arguments{
\item{datasetName}{Name of the dataset. (e.g. "S.0.1.Rsq.0.1").
Usually, it is has the same name as \code{basename(top.dir)}.}

\item{toolName}{Name of computational approach. (e.g. "SigProExtractor")}

\item{tool.dir}{Fourth level path from the \code{top.dir}. Expected to have
multiple runs with different names (e.g. "seed.1")
That is,
\code{top.dir}/sp.sp/ExtrAttr/sa.results/. or
\code{top.dir}/sa.sa.96/Attr/deconstructSigs.results/

Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{tool.dir}
which stores the software output.}

\item{run.names}{A character vector records the list of \code{run.dir},
or fifth level directories from the dataset top-level folder.
E.g., c("seed.1","seed.691")}
}
\value{
A list contain values of measures measures in multiple runs: \itemize{
\item $averCosSim Cosine similarity
\item $truePos True Positives(TP): Ground-truth signatures which are active in
the spectra, and extracted.
\item $falseNeg False Negatives(FN): Ground-truth signatures not extracted.
\item $falsePos False Positives(FP): Signatures wrongly extracted, not resembling
any ground-truth signatures.
\item $TPR True positive rate (TPR, Sensitivity): TP / (TP + FN)
\item $PPV Positive predictive value (PPV): TP / (FP + TP)
\item $cosSim Average cosine similarity to each of the ground-truth signatures.
\item $ManhattanDist Scaled Manhattan distance between ground-truth and inferred
exposures to each of the ground-truth signatures.
}
This list also contains \code{mean} and \code{sd}, and other
statistics of these measures in \itemize{
\item $fivenum
\item $fivenumMD
\item $meanSD
\item $meanSDMD
}
}
\description{
Summarize results from each computational approach in \code{tool.dir}/\code{run.names}
(generated by running a computational approach),
combine them into \code{tool.dir}.
}
\details{
Also writes multiple files into folder \code{tool.dir}.
}
