% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeMultiToolsMultiDatasets}
\alias{SummarizeMultiToolsMultiDatasets}
\title{Summarize results for multiple datasets, by different computational approaches.}
\usage{
SummarizeMultiToolsMultiDatasets(
  dataset.dirs,
  second.third.level.dirname,
  out.dir,
  overwrite = FALSE
)
}
\arguments{
\item{dataset.dirs}{Paths of top-level dataset directories trees you want
to investigate.
E.g. "./S.0.1.Rsq.0.1"}

\item{second.third.level.dirname}{Name of the second.level.dir (e.g. "sp.sp")
and the third.level.dir (e.g. "ExtrAttr") to be investigated.

Examples are: "sp.sp/ExtrAttr", "sa.sa.96/Attr"

Note: \code{multiTools.RDa} are expected to be exist under
\code{dataset.dirs}/\code{second.third.level.dirname}.}

\item{out.dir}{Path of the output directory.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results of mutational signature extraction and exposure inferrence
by multiple computational approaches on multiple datasets. Before running this
function, make sure the summary file for each single data set
 \code{third.level.dir}/\code{multiTools.Rda} exists.
}
\details{
\code{multiTools.Rda} is generated by \code{\link{SummarizeMultiToolsOneDataset}}).
}
