% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeOneToolMultiDatasets.R
\name{SummarizeOneToolMultiDatasets}
\alias{SummarizeOneToolMultiDatasets}
\title{Combine results for multiple datasets, from one computational approaches.}
\usage{
SummarizeOneToolMultiDatasets(
  datasetNames = SynSigGen::SBS1SBS5datasetNames,
  datasetGroup,
  datasetGroupName,
  datasetSubGroup = NULL,
  datasetSubGroupName = NULL,
  toolName,
  toolPath,
  out.dir,
  display.datasetName = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{datasetNames}{Names of datasets which are also folder names
under \code{toolPath}. These folders contain results of \code{toolName}
on such datasets.
E.g. \code{SynSigGen::SBS1SBS5datasetNames}}

\item{datasetGroup}{Numeric or character vector differentiating
datasets within each group.
E.g. For SBS1-SBS5 correlated datasets, we can consider the value of
SBS1-SBS5 exposure ratio as the value for \code{datsetgroup}:
\code{rep(c(0.1,0.5,1,2,5,10),each = 4)}

The value is set to \code{Default} if unspecified.}

\item{datasetGroupName}{Meaning of all datasetGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider
\code{"SBS1-SBS5 exposure ratio"}
as what \code{datasetGroup} is referring to.}

\item{datasetSubGroup}{Numeric or character vector differentiating
datasets within each sub-group.
E.g. For SBS1-SBS5 correlated datasets, we can consider the value of
\code{SBS1-SBS5 correlation} as the value of \code{subgroup}:
\code{rep(c(0.1,0.2,0.3,0.6),times = 5)}}

\item{datasetSubGroupName}{Meaning of all \code{datasetSubGroup}.
E.g. For SBS1-SBS5 correlated datasets, we can consider
\code{"SBS1-SBS5 correlation"}
as what \code{datasetSubGroup} is referring to.}

\item{toolName}{Name of computational approach to be investigated
(e.g. "SigProExtractor")}

\item{toolPath}{The path of the results of the computational approach
to be investigated. May include top-level directory
(e.g. \code{3a.Original_output_K_unspecified}) and second-level
directory containing outputs and summaries of one computational approach
to be investigated (e.g. \code{SigProExtractor} or \code{SigProExtractor.results}).

One example: \code{3a.Original_output_K_unspecified/SigProExtractor}

Note: this function expects file \code{multiRun.RDa} generated by
\code{SummarizeMultiRuns} under \code{toolPath}/\code{datasetNames}}

\item{out.dir}{Path of the output directory.}

\item{display.datasetName}{Whether to put the name of spectra datasets inside of
the csv outputs of summary tables.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results from each computational approach in \code{toolPath}/\code{datasetNames}
and combine them into \code{out.dir}.
}
